+++ packages/anima-cli/src/api/index.ts
@@ -1,3 +1,2 @@
 export * from './authenticate';
-export * from './storybook';
 export * from './stories';
