+++ packages/storybook-api/src/index.ts
@@ -1,0 +1,2 @@
+export { StorybookApi } from "./storybook";
+export type * from "./types";
